Add files via upload
This code is a Lua script that is designed to update or install certain files from a specified branch of a GitHub repository. The script first checks if the device has access to the internet by checking if the "http" library is present. If not, the script will print "No access to web" and exit. The branch of the GitHub repository is specified as "master" at the beginning of the script. The script then defines an array of files to be updated or installed, with each file having a name and a URL that is constructed by concatenating the branch and file name.

The script then iterates through the array of files, and for each file, it checks if the file already exists on the device and if so, it reads the text of the file. The script then prints a message indicating whether it will update or install the file, and where it will be installed. The script then uses the "http" library to send a GET request to the URL of the file, and if the request is successful, it checks the HTTP response code. If the code is 200, indicating a successful response, the script reads the text of the file from the website, and compares it to the text of the file on the device. If the texts match, the script prints "skip", otherwise it writes the text from the website to the file on the device and prints "done". If the response code is not 200, the script prints an error message. After updating or installing each file, the script sleeps for 0.1 seconds before moving on to the next file. The script ends by printing "Finished".